requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
httpx
apscheduler
twilio
google-generativeai
python-dotenv
pytest
pytest-asyncio
